fix: given when then 반영, 패키지 변경
